changed int type to long

+++ test/DataTypesTest.java
@@ -20,10 +20,12 @@ public class DataTypesTest {
         // Note: You can use _ to help make it easier to
         //       read large numbers, as is done below.
         long x = 500_000_500_000L;
-        List<Integer> lst = new ArrayList<>();
-        for (int i = 1; i <= 1_000_000; i++) {
+        List<Long> lst = new ArrayList<>();
+        for (long i = 1; i <= 1_000_000; i++) {
             lst.add(i);
         }
         assertEquals("sum form 1 to 1 million should be " + x, x, DataTypes.sum(lst));
     }
+
+
 }
